scripts/requirements.txt
mysql-connector-python
progressbar
validate_email
